Do not automatically invalidate resources if the fetch action throws/rejects
This is a breaking change. However it gives users of this library more control over what happens if an error occurs in a resource. Instead of forcefully invalidating the resource, errors are now treated the same way as normal return values. The default behaviour is that errors are kept as long there is a transaction that is using the resource. This behaviour can be tweaked by using the new "maximumRejectedStaleness" setting, which works the same as "maximumStaleness", however it only has an effect if the resource instance had thrown an error or returned a rejected promise.

To get the old behaviour use:

```js
manager.resource({
  name: 'food',
  fetch: async (params, {invalidate}) => {
    try {
      await foodFromApi(params);
    }
    catch(err) {
      invalidate();
      throw err;
    }
  },
})
```